update: melhorias nas classes e nos testes do projeto add: inclusao do HeroControllerIT todo: descobrir por que ele nao roda

+++ interview-junior-OG/application/src/main/java/br/com/gubee/interview/core/application/powerstats/commands/DeletePowerStatsCommand.java
@@ -5,7 +5,6 @@ import br.com.gubee.interview.core.application.ports.repositories.PowerStatsRepo
 import lombok.RequiredArgsConstructor;
 
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.UUID;
 
@@ -14,8 +13,16 @@ import java.util.UUID;
 public class DeletePowerStatsCommand {
     private final PowerStatsRepository powerStatsRepository;
 
-    @Transactional
-    public void execute(UUID id) {
+    public boolean execute(UUID id) {
+        if (id == null) {
+            throw new IllegalArgumentException("PowerStats ID cannot be null");
+        }
+
+        if (powerStatsRepository.findById(id).isEmpty()) {
+            return false;
+        }
+
         powerStatsRepository.delete(id);
+        return true;
     }
 }
